hecho

+++ cpp09/ex01/inc/RPN.hpp
@@ -6,7 +6,7 @@
 /*   By: vpeinado <victor@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/08/22 20:17:53 by vpeinado          #+#    #+#             */
-/*   Updated: 2024/08/22 21:54:14 by vpeinado         ###   ########.fr       */
+/*   Updated: 2024/08/22 22:36:29 by vpeinado         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,7 +19,9 @@
 class RPN
 {
     private:
-        std::stack<double> _values;
+        //Alternando double a int, podemos hacer que la calculadora RPN funcione con números enteros o decimales.
+        //std::stack<double> _values;
+        std::stack<int> _values;
         std::string _str;
         RPN();
     public:
